analyzer comp

+++ client_bak/src/java_rules_of_the_grammer.h
@@ -104,7 +104,7 @@ struct classdescinfo {
 };
 
 struct newhandle {
-    unsigned short handle;
+    unsigned int handle;
 };
 
 struct newclassdesc {
